% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/non_paramteric.R
\name{non_parametric}
\alias{non_parametric}
\title{Non Parametric Estimate}
\usage{
non_parametric(data, zij, n, p)
}
\arguments{
\item{data}{The whole dataset containing pairwise communications, the first column
is the person with outgoing communication and the second column is the person with
incoming communication. The last column is the communication time point (should be 
in numeric format).}

\item{zij}{The transfromed non-homogeneous effect (array with dimension (n,n,p)). The 
transformation can be done using z_projection() function.}

\item{n}{number of individuals in the whole network.}

\item{p}{number of non-homogeneuous effect.}
}
\value{
\item{homo_coefficients}{List Individual's cumulative effect from scaled time range 
[0,1]. The parameters are estimated at (0.01, 0.02, ..., 1)}
\item{homo_sd}{List Individual's standard deviation for cumulative effect from scaled 
time range [0,1]. The parameters are estimated at (0.01, 0.02, ..., 1)}
\item{nonhomo_coefficients}{List nonhomogeneous effect from scaled time range 
[0,1]. The parameters are estimated at (0.01, 0.02, ..., 1)}
\item{nonhomo_sd}{List nonhomogeneous parameter's standard deviation for cumulative 
effect from scaled time range [0,1]. The parameters are estimated at (0.01, 0.02, ..., 1)}
}
\description{
This function allows you to do non parametric estimation for additive network model
}
