% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proxyData.R
\docType{data}
\name{proxyData}
\alias{proxyData}
\title{MIT Social network daat}
\format{
An object of class \code{list} of length 8.
}
\usage{
data(proxyData)
}
\description{
MIT Social network daat
}
\examples{
data(proxyData)
}
\keyword{datasets}
