% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xConstruct.R
\name{xConstruct}
\alias{xConstruct}
\title{Projection and X creation}
\usage{
xConstruct(n, A = NULL)
}
\arguments{
\item{n}{Sample size}

\item{A}{The non-homogeneous effect (array with dimension (n,n,p))}
}
\value{
\item{intercept}{Intercept}
\item{x}{Homofily effect}
\item{zij}{The transfromed non-homogeneous effect (array with dimension (n,n,p))}
}
\description{
Create X matrix
}
