% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zProjection.R
\name{zProjection}
\alias{zProjection}
\title{Projection of Non-homogeneous effect}
\usage{
zProjection(A)
}
\arguments{
\item{A}{The non-homogeneous effect (array with dimension (n,n,p))}
}
\value{
\item{zij}{The transfromed non-homogeneous effect (array with dimension (n,n,p))}
}
\description{
Construct the non-honogeneous effect on residual space (Proporsition 1)
}
