% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paramteric.R
\name{parametric}
\alias{parametric}
\title{Semi Parametric Estimate}
\usage{
parametric(data, zij, n, p, tz, k = 0)
}
\arguments{
\item{data}{The whole dataset containing pairwise communications, the first column
is the person with outgoing communication and the second column is the person with
incoming communication. The last column is the communication time point (should be
in numeric format).}

\item{zij}{The transfromed non-homogeneous effect (array with dimension (n,n,p)). The
transformation can be done using zProjection() function.}

\item{n}{number of individuals in the whole network.}

\item{p}{number of non-homogeneuous effect.}

\item{tz}{Time points for non-homofily effects (ranging from (0,1)), the length of this
vector should equal to the 4th dimension of zij which reflect the non-homofily effect
on or before those specific time points.}

\item{k}{number of orthogonal time dependent covariates}
}
\value{
\item{homo_coef}{List Individual's parameters' estimation}
\item{homo_cov}{List Individual's parameters' covairance matrix}
\item{nonhomo_coef}{List nonhomogeneous parameters' estimation}
\item{nonhomo_cov}{List nonhomogeneous parameter's covariance estimation}
}
\description{
This function allows you to do non parametric estimation for additive network model.
For non-homofily effect, we only consider non-time dependent covariates since it is
the most common case in real world while time dependent cases are rare and difficult
to be specified. Also, since the projection should be done at any time t, the model
with time dependent covariates requires much higher computation power.
}
