% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paramteric.R
\name{parametric}
\alias{parametric}
\title{Semi Parametric Estimate}
\usage{
parametric(data, zij, n, p, k = 0)
}
\arguments{
\item{data}{The whole dataset containing pairwise communications, the first column
is the person with outgoing communication and the second column is the person with
incoming communication. The last column is the communication time point (should be
in numeric format).}

\item{zij}{The transfromed non-homogeneous effect (array with dimension (n,n,p)). The
transformation can be done using z_projection() function.}

\item{n}{number of individuals in the whole network.}

\item{p}{number of non-homogeneuous effect.}

\item{k}{number of orthogonal time dependent covariates}
}
\value{
\item{homo_coef}{List Individual's parameters' estimation}
\item{homo_cov}{List Individual's parameters' covairance matrix}
\item{nonhomo_coef}{List nonhomogeneous parameters' estimation}
\item{nonhomo_sd}{List nonhomogeneous parameter's standard deviation}
}
\description{
This function allows you to do non parametric estimation for additive network model
}
